requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
discord.py==2.3.1
exceptiongroup==1.1.2
fonttools==4.39.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.7.1
moviepy==1.0.3
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
openai==0.27.8
opencv-python==4.7.0.68
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
proglog==0.1.10
pycparser==2.21
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2022.7.1
pywin32==305
redis==4.6.0
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.1
scipy==1.10.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
spotipy==2.23.0
threadpoolctl==3.1.0
tqdm==4.64.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.14
wsproto==1.2.0
yarl==1.9.2
